Move reaction list and buttons into their own components
